rename masked_labeled_image
