+++ ipp_toolkit/data/masked_labeled_image.py
