+++ ipp_toolkit/experiments/missions.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from ipp_toolkit.config import MEAN_ERROR_KEY, VIS_LEVEL_3
-from ipp_toolkit.data.MaskedLabeledImage import MaskedLabeledImage
+from ipp_toolkit.data.masked_labeled_image import MaskedLabeledImage
 from ipp_toolkit.planners.masked_planner import BaseGriddedPlanner
 from ipp_toolkit.predictors.intrestingness_computers import BaseInterestingessComputer
 from ipp_toolkit.predictors.masked_image_predictor import MaskedLabeledImagePredictor
